+++ CHANGELOG.md
@@ -15,8 +15,8 @@
 * Added more logging of errors
     * Added the `showConsoleErrors` setting to hide errors from the Symfony console
 * Added progress indicator on the status bar
-* [#1](https://github.com/TheNouillet/symfony-vscode/issues/1) Added buttons to the side of TreeViews to re-sync the extension and Symfony.
-* [#3](https://github.com/TheNouillet/symfony-vscode/issues/3) Added class name for services aliases
+* Added buttons to the side of TreeViews to re-sync the extension and Symfony.
+* Added class name for services aliases
 
 ## 0.0.1