+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Added aucompletion of public services in PHP files
 * Added the `enableFileWatching` setting to enable or disable file watching.
+* Added the `servicesFilters` to improve autocompletion pertinence.
 * Added the "Toggle class/id display for services" command to switch between Id and class name display on the services view.
 * Added the "Toggle path/id display for routes" command to switch between Id and paths display on the routes view.
     * These two commands are available via buttons on the side of the two views.